# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AwakeMovies Scraper API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScrapeSource {
  id: string;
  label: string;
  needsSite: boolean;
}

/**
 * Whether to scrape as a movie or a series
 */
export type ScrapeRequestMode = typeof ScrapeRequestMode[keyof typeof ScrapeRequestMode];


export const ScrapeRequestMode = {
  movie: 'movie',
  series: 'series',
} as const;

export interface ScrapeRequest {
  /** URL of the movie or series page to scrape */
  url: string;
  /** Whether to scrape as a movie or a series */
  mode: ScrapeRequestMode;
}

/**
 * Whether to scrape as a movie or a series
 */
export type ScrapeNkiriRequestMode = typeof ScrapeNkiriRequestMode[keyof typeof ScrapeNkiriRequestMode];


export const ScrapeNkiriRequestMode = {
  movie: 'movie',
  series: 'series',
} as const;

/**
 * Which site variant to use (auto-detected from URL if omitted)
 */
export type ScrapeNkiriRequestSite = typeof ScrapeNkiriRequestSite[keyof typeof ScrapeNkiriRequestSite];


export const ScrapeNkiriRequestSite = {
  nkiri: 'nkiri',
  dramakey: 'dramakey',
} as const;

export interface ScrapeNkiriRequest {
  /** URL of the movie or series page to scrape */
  url: string;
  /** Whether to scrape as a movie or a series */
  mode: ScrapeNkiriRequestMode;
  /** Which site variant to use (auto-detected from URL if omitted) */
  site?: ScrapeNkiriRequestSite;
}

export interface DownloadLink {
  quality: string;
  link: string;
}

export interface Episode {
  episode: number;
  title: string;
  url: string;
}

export type _ScrapeResultAwptAdditionalInfo = { [key: string]: unknown };

export interface ScrapeResult {
  error?: string;
  _awpt_type?: string;
  _awpt_title?: string;
  _awpt_synopsis?: string;
  _awpt_poster?: string;
  _awpt_trailer_url?: string;
  _awpt_genre?: string;
  _awpt_stars?: string;
  _awpt_release_date?: string;
  _awpt_country?: string;
  _awpt_ratings?: string;
  _awpt_language?: string;
  _awpt_subtitles?: string;
  _awpt_source?: string;
  _awpt_duration?: string;
  _awpt_file_size?: string;
  _awpt_status?: string;
  _awpt_season?: string;
  _awpt_total_episodes?: string;
  _awpt_download_link?: DownloadLink[];
  _awpt_episodes?: Episode[];
  _awpt_additional_info?: _ScrapeResultAwptAdditionalInfo;
  [key: string]: unknown;
 }

export interface ErrorResponse {
  error: string;
}