# lib/api-zod/src/generated/types/downloadLink.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AwakeMovies Scraper API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadLink {
  quality: string;
  link: string;
}