# lib/api-zod/src/generated/types/scrapeNkiriRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AwakeMovies Scraper API
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapeNkiriRequestMode } from './scrapeNkiriRequestMode';
import type { ScrapeNkiriRequestSite } from './scrapeNkiriRequestSite';

export interface ScrapeNkiriRequest {
  /** URL of the movie or series page to scrape */
  url: string;
  /** Whether to scrape as a movie or a series */
  mode: ScrapeNkiriRequestMode;
  /** Which site variant to use (auto-detected from URL if omitted) */
  site?: ScrapeNkiriRequestSite;
}