# lib/api-zod/src/generated/types/scrapeSource.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AwakeMovies Scraper API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeSource {
  id: string;
  label: string;
  needsSite: boolean;
}